Use of switch case instead of if-else and seperate file for export
