+++ app.js
@@ -3,6 +3,7 @@ import cartogram from "./catogram"
 import { getRadius, getGridData, getTransformation, getPath } from './shaper';
 import { mover, mout, mclickBase, mclick, dragstarted, dragged, dragended } from './mouse-events';
 import { colors, margin, width, height, strokeWidth } from './constants';
+import { download } from './export';
 
 var exportJson
 
@@ -33,16 +34,7 @@ yearButton.addEventListener('click', () => {
 
 downloadButton.addEventListener('click', () => {
   let fileType = document.querySelector('#download-option').value;
-  var filename = "cartogram" + yearInput.value;
-  if (fileType == "Geojson") {
-    downloadObjectAsJson(exportJson, filename)
-  } else if (fileType == "SVG") {
-    d3.select("#download").each(function () {
-      d3.select(this)
-        .attr("href", 'data:application/octet-stream;base64,' + btoa(d3.select("#container").html()))
-        .attr("download", filename + ".svg")
-    })
-  }
+  download(fileType, yearInput.value);
 });
 
 cellShapeButton.addEventListener('click', () => {
@@ -194,16 +186,6 @@ function getData(data) {
     obj[data[x].code] = data[x]
   }
   return obj
-}
-
-function downloadObjectAsJson(exportObj, exportName) {
-  var dataStr = "data:text/json;charset=utf-8," + encodeURIComponent(JSON.stringify(exportObj));
-  var downloadAnchorNode = document.createElement('a');
-  downloadAnchorNode.setAttribute("href", dataStr);
-  downloadAnchorNode.setAttribute("download", exportName + ".json");
-  document.body.appendChild(downloadAnchorNode);
-  downloadAnchorNode.click();
-  downloadAnchorNode.remove();
 }
 
 start()
