+++ index.html
@@ -21,7 +21,7 @@
     </div>
 
     <input class="left form-control form-control-sm round-right" style="width: 50px; margin-left: 40px;" type="number"
-      id="radius" name="radius" min="1" max="12" value="4" />
+      id="radius" name="radius" min="1" max="12" value="5" />
     <input type="button" value="Select Radius" class="left btn btn-sm btn-success round-left"
       id="select-radius"></input>
 
@@ -40,8 +40,8 @@
 
     <select id="download-option" class="left form-select form-select-sm round-right"
       style="width: 100px; margin-left: 10px;">
-      <option>Geojson</option>
       <option>SVG</option>
+      <option>GeoJSON</option>
     </select>
     <button class="btn btn-sm btn-secondary left round-left"><a id="download" href="#">Download</button>
